Move files from src/main/parsers/* to src/parsers/*
