+++ src/common/Facade.ts
@@ -1,17 +1,16 @@
-import * as ParserFactory from "../main/parsers/ParserFactory";
+import * as ParserFactory from "./../parsers/ParserFactory";
 import { ClassMetadata as ClassType, FileMetadata } from "../common/types/backend.type";
-import { FileFactory, Workspace } from "../main/util";
+import { FileFactory } from "./../main/util";
 
 const CachFile: { [key: string]: ClassType } = { }
 
 export const ClassMetadata = {
-	fromPackageParts: (parts: string[]): undefined | ClassType => {
-		const absolutePath = Workspace.getAbsolutePath(parts);
-		return absolutePath == undefined ? undefined : ClassMetadata.fromAbsolutePath(absolutePath);
-	},
 
 	fromAbsolutePath: (absolutePath: string): undefined | ClassType => {
 		const file = FileFactory.fromAbsolutePath(absolutePath);
+		if (file == undefined) {
+			return undefined;
+		}
 		return ClassMetadata.fromFileMetadata(file);
 	},
