+++ src/parsers/ParserFile.ts
@@ -1,8 +1,8 @@
-import { ClassMetadata, FileMetadata } from "../../common/types/backend.type";
-import { Optional } from "../../common/types/classes.type";
-import { IParserFile } from "../../common/types/interfaces.type";
-import { Mock } from "../../common/types/mock.types";
-import { FileReader } from "../util";
+import { ClassMetadata, FileMetadata } from "../common/types/backend.type";
+import { Optional } from "../common/types/classes.type";
+import { IParserFile } from "../common/types/interfaces.type";
+import { Mock } from "../common/types/mock.types";
+import { FileReader } from "../main/util";
 import { RegexGroups, ParserFileRegex } from "./ParserFileRegex";
