+++ src/parsers/java/JavaImportParser.ts
@@ -1,7 +1,7 @@
-import { Package } from "../../../common/types/backend.type";
-import { Optional } from "../../../common/types/classes.type";
-import { IPackageMapper, IParser } from "../../../common/types/interfaces.type";
-import { Regex } from "../../util";
+import { Package } from "../../common/types/backend.type";
+import { Optional } from "../../common/types/classes.type";
+import { IPackageMapper, IParser } from "../../common/types/interfaces.type";
+import { Regex } from "../../main/util";
 import { RegexGroups } from "../ParserFileRegex";
 
 enum Attrs {
