+++ src/parsers/java/JavaPackageMapper.ts
@@ -1,7 +1,7 @@
-import { FileMetadata } from "../../../common/types/backend.type";
-import { WorkspaceFiles } from "../../../common/types/classes.type";
-import { IPackageMapper } from "../../../common/types/interfaces.type";
-import { Workspace } from "../../util";
+import { FileMetadata } from "../../common/types/backend.type";
+import { WorkspaceFiles } from "../../common/types/classes.type";
+import { IPackageMapper } from "../../common/types/interfaces.type";
+import { Workspace } from "../../main/util";
 
 export class JavaPackageMapper implements IPackageMapper {
