+++ src/parsers/php/PhpNamespaceParser.ts
@@ -1,7 +1,7 @@
-import { Namespace } from "../../../common/types/backend.type";
-import { Optional } from "../../../common/types/classes.type";
-import { IParser } from "../../../common/types/interfaces.type";
-import { Regex } from "../../util";
+import { Namespace } from "../../common/types/backend.type";
+import { Optional } from "../../common/types/classes.type";
+import { IParser } from "../../common/types/interfaces.type";
+import { Regex } from "../../main/util";
 import { RegexGroups } from "../ParserFileRegex";
 
 enum Def {
